feat: Create `view-wrapper` to allow view change
`view-wrapper` is a Reagent component that just subscribes to the `view`
state and renders the required view component.

This will allow us to easily change views on user interaction, which should
appear to the user that they are visiting a new page e.g. we have created an
example of this with the buttons on our `propery` and `portfolio` views.

To improve the user experience we should try and override the `back` button
on the user's browser to update the view state to the previous state. This will
appear to the user that they are going back to the previous "page", which is
a far better user experience.